pyproject.toml
[tool.poetry]
name = "brain-ag"
version = "0.1.0"
description = ""
authors = ["Felipe Silva <felipefets@gmail.com>"]
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-django = "^4.11.1"
django-environ = "^0.12.0"
psycopg2-binary = "^2.9.10"
django = "^5.2.4"
djangorestframework = "^3.16.0"
validate-docbr = "^1.11.1"
django-filter = "^25.1"
markdown = "^3.8.2"
drf-spectacular = "^0.28.0"
redis = "^6.2.0"
django-redis = "^6.0.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
asyncio_mode = "auto"
addopts = "-p no:warnings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
